+++ graph-http/src/download/mod.rs
@@ -1,3 +1,5 @@
 mod client;
+mod error;
 
 pub use client::*;
+pub use error::{AsyncDownloadError, BlockingDownloadError};
